Upgrade pubsub samples.
