+++ pubsub/cloud-client/requirements.txt
@@ -1,1 +1,1 @@
-gcloud==0.18.3
+google-cloud==0.19.0
