+++ pubsub/cloud-client/topics_test.py
@@ -16,9 +16,9 @@ from gcloud import pubsub
 from gcp.testing import eventually_consistent
 import pytest
 
-import publisher
+import topics
 
-TEST_TOPIC = 'publisher-test-topic'
+TEST_TOPIC = 'topics-test-topic'
 
 
 @pytest.fixture
@@ -30,38 +30,62 @@ def test_topic():
         topic.delete()
 
 
-def test_list(test_topic, capsys):
+def test_list_topics(test_topic, capsys):
     test_topic.create()
 
     @eventually_consistent.call
     def _():
-        publisher.list_topics()
+        topics.list_topics()
         out, _ = capsys.readouterr()
         assert test_topic.name in out
 
 
-def test_create(test_topic):
-    publisher.create_topic(test_topic.name)
+def test_create_topic(test_topic):
+    topics.create_topic(test_topic.name)
 
     @eventually_consistent.call
     def _():
         assert test_topic.exists()
 
 
-def test_delete(test_topic):
+def test_delete_topic(test_topic):
     test_topic.create()
 
-    publisher.delete_topic(test_topic.name)
+    topics.delete_topic(test_topic.name)
 
     @eventually_consistent.call
     def _():
         assert not test_topic.exists()
 
 
-def test_publish(test_topic, capsys):
+def test_publish_message(test_topic, capsys):
     test_topic.create()
 
-    publisher.publish_message(test_topic.name, 'hello')
+    topics.publish_message(test_topic.name, 'hello')
 
     out, _ = capsys.readouterr()
     assert 'published' in out
+
+
+def test_get_topic_policy(test_topic, capsys):
+    topics.get_topic_policy(test_topic.name)
+
+    out, _ = capsys.readouterr()
+    assert test_topic.name in out
+
+
+def test_set_topic_policy(test_topic):
+    topics.set_topic_policy(test_topic.name)
+
+    policy = test_topic.get_iam_policy()
+    assert policy.viewers
+    assert policy.editors
+
+
+def test_test_topic_permissions(test_topic, capsys):
+    topics.test_topic_permissions(test_topic.name)
+
+    out, _ = capsys.readouterr()
+
+    assert test_topic.name in out
+    assert 'pubsub.topics.publish' in out
